Finer date control
Added support for handling starting and ending day and not just
year and month, as part of the proc_days logic.

Inturn added a helper routine to convert a date string into internal
date dictionary format. The same is inturn used now to process the
commandline start and end user arguments, ie if the user runs the
program in the non interactive mode. So user has finer date control
now.