Restructure the flow for easy use
The logic for fetching data for a given date range, moved into a
function of its own.

The same is now called by the program during the non interactive
mode, as well as can be called by a user in interactive mode without
having to worry about creating the internal date dictionaries wrt
start and end.

Now the user can just pass the date as a string in YYYY[MM[DD]]
format.