Load fetched data, csv fname template, name set
Now the global data dictionary maintains the name in a set, so that
we can know if the fetched data has the same MFCode assigned to more
than 1 MF and or if there are missplet MFs across the fetched data
sets(CSVs).

The filename and path used wrt the fetched data is maintained as
part of a template string, so that the same can be used by the
fetch as well as the load logic to identify the file to work with
for any given date.

Added helper routines for loading fetched data for a given date
or a range of dates. This uses the parse_csv logic previously
implemented.