Make findmatchingmf flexible
One can either search for a exact match of the given full string,
or try to match based on the tokens in the given string (which
inturn could be fully or partially matched at the token level).

The user can also control whether the match should ignore the case
of the string contents or not.