proc_days takes NotBeyondYesterday as a argument
Instead of directly working with the global gbNotBeyondYesterday
variable, now proc_days takes it as a seperate argument to itself.

So users of proc_days can decide whether to limit it to data till
yesterday or to let it go beyond it.

Also as day info need not be part of the internal date dictionary
format, so avoid the check and adjusting of day  at the begining
of the proc_days logic. EItherway there is a seperate logic to filter
based on the day being processed in the inner loop of proc_days
logic.