lookatmfs: handle case where data starts inbtw date range
Different MFs will have different start dates (i.e when they were
started), so when looking at MFs for relative study, we require
to pick the start reference nav value, as that of the date within
the date range, when the MF actually started (i.e for MFs, which
didnt start on or before the start of the date range being studied).

And obviously for those MFs, which had started well before the start
of the date range being studied, we take the starting date of the
date range, as the date for its reference NAV value for relative
studies.