breakpoint and realise the mistake wrt argsort
Added breakpoint to get into the nuts and bolts of the prg execution,
so that I could identify the strange sorting issue I was noticing.

Inturn cross verifying the strangeness from within the debugger,
relooked into argsort once again and realised that what it returns
is the array, which at position gives the index into the original
non sorted array that should be used to get data into that position
in a sorted array. While I had wrongly assumed that each position
in the returned array gives the position in the sorted array into
which the element in that position of the non sorted array should
be placed.

ie
it returns argsortResult =>
[indexFromNonSortedArray4ThisPosInSortedArray, ...]

and not argsortResult =>
[indexIntoSortedArray4ThisPosInNonSortedArray, ...]

Need to update the logic accordingly next.

NOTE: The issues with jumping between different domains and languages,
... and then coming back and not refreshing the documentation properly
before starting on some new stuff using existing things (inc langs) ;)