lookup_data and Cleanup usage flow related logic
Have renamed
	lookupmfs to lookupmfs_names.
	look4mfs to lookupmfs_ops

Added a new lookup_data, which depending on whether a string
or list is passed as the first argument will either call into
lookupmfs_ops or lookupmfs_names.

lookupmfs_names can now work with either a list of strings or
a single string with ';' as seperator between independent set
of name match tokens.

Updated the usage note as well as non interactive usage flow to
mirror the above and last few updates.