Rope in readline, curl no use back to wget
Imported readline, so that user has the power of readline, when
entering commands like recalling(up/down arrows, editing(left/right
arrows) etc.

We need a way of trying to find if there is a new file on the server
or not and inturn update local file only if remote file is newer.
Wget doesnt support this if we use custom output filenaming.
Tried using the time-condition of curl, but doesnt seem to help.
So going back to wget again.

Some good entertainment which kept me company over last few days,
when indulging in this and few other experiments.

	Agatha Cristie's Poirot
	The Planets
	Swapna Sanchari
	Palunku