If prev downloaded file redownloaded, then get fresh
Because of the way wget is used here, where we explicitly set
the dest file to write to, it doesnt support using timestamps
to identify if local and remote files are same or not.

Parallely, the script on the server which is generating / returning
the requested nav data, doesnt seem to maintain timestamp in a
proper way. So eitherway it wouldnt have helped.

Instead we depend on wget fetching data new only if there is more
data wrt a given file on the server. However in this case it assumes
that the local file is only partial and so it appends to it.

But wrt NAV data, what is noticed is that the server may send a
new version of the file with added nav value or so for the prev
day or two, if there was a delay in updating the same from the
AMC/Fund house. So in this case, we need to refetch the data,
rather than appending to existing file (which will mess it up
with partly old and partly new data).

So added logic to remove and refetch file, if wget has redownloaded
a file.

NOTE: If the updated file is smaller, then the overall logic may
miss it, as the local file is larger, have to check wgets default
behaviour in such a case. The only way to overcome this perfectly,
in the worst case will be to refetch each and every file. And or
user needing to remove such files (based on info from outside this
program) and then the program automatically refetch it.