TabComplete only if user is entering first word
Check for num of tokens < 1, means user has not yet entered anything

Check for numTokens == 1, would trigger when either user has entered
one full token or the is in the process of entering the 1st token.

It will also trigger if user is about to start on entering the 2nd
token, i.e by putting a space after the 1st token.

The check for text != "" ensures that the logic doesnt trigger if
user as conveyed his intention to enter the 2nd token by placing a
space after the first token.