Initial skeleton for a new procdata_ex logic
This will allow one to operate on the data such that one can
specify what data to operate on and inturn save the result into
gData dictionary, so that it can be reused later if required in
other operations.

Also it tries to operate on things in a bit more numpy way.

Only initial skeleton added for now.