World: fetch data about some world market datasets
Found a source for some of the world datasets at

	https://fred.stlouisfed.org/series/

This seems to provide key datasets like US market indexes, gold,
crude, silver.

NOTE: If the date given is invalid or is for a date earlier than
what the server is configured for, it seems to dump the full
dataset in one go.

NOTE: For few market indexes it seems to be limited to 10Yrs.