Day of reckoning ;-)
Have fixed a oversight when implementing the new findmatchingmf,
which depends on generified original logic now called _findmatching
internally. The _findmatching logic cant try to map from index to
code, as the list of strings(i.e the dataSet) passed to it could
be anything and not necessarily the list of MFs. Both had to be
updated to account for things suitably.

Also ReRealised that my original findmatchingmf which inturn is now
called _findmatching, actually take a searchTemplate and checks if
any of the names in a list given to it, match the given template
or not.

TODO: This also means that I cant use the _findmatching wrt the new
white and blacklists wrt MFTypes and MFNames. Need to implement a
new matches_templates logic, which inturn can be used by these,
as well as by _findmatching hopefully.