Identify oversight wrt findmatchingmf to _findmatching
When the original findmatchingmf logic was renamed into _findmatching
with a generic semantic, and inturn findmatchingmf was built on top
of it, I had missed out on handling the index2code properly.

Inturn yesterday or today early morning, I had fixed the oversight
by moving the handling of index2code into findmatchingmf from
_findmatching. However in the process, I had forgotten to use the
right list, when doing so.

THis has been fixed now.