Transition parse_csv to matches_templates
As _findmatching checks a match template against a list of strings
to see if any of the strings match the template given, this cant
be used within parse_csv to see if a given string matches any of
the given set of match templates.

So matches_templates was implemented to provide this generic func.
ANd inturn parse_csv updated to call the same now, as already noted
as required yesterday afternoon. This completes the basic skeleton
wrt that todo now.