Fine tune global load filters
As the equity related MFs could be part of other MFTypes, the
same has been updated into the WhiteListMFTypes.

Similarly as - could be prefixed before Dividend or Direct or so
so now the filtering out of dividend scheme is using a partial match
for Dividend token.

TODO: May add logic to cleanup names by replacing '-' with ' '