_show_plot: show_plot split into _show_plot and itself
So when user is plotting a series of data, which all belong to
the same date range, independent of one another, there is no need
to clear the dateRange. So added _show_plot for use in this situation.

TODO: Rather independent of above, There should be the base level
dateRange, which maps to full data dateRange. ANd then any custom
dateRange set by user using lookat (rather a new version of it).
And show_plot should clear only the dateRange set of lookat and
not the base level dateRange.