AnalDataSimple: Only work with currently active entities
THe logic checks if the entities to work with had data associated
with them loaded wrt a date atleast 1 weak from last date in the
loaded data set. If not, entity will be ignored wrt the operation.

Also ensure that theSaneArray used is a copy and not a reference
to the original dataset.