AnalDataSimple: Check MinEntityLifeDataInYears always
Previously the logic would check for a minimum amount of data
wrt each entity being ranked, only if the operation type was
srel related.

Instead NOW, the logic will check for the availability of
specified amount of minimum historical data, always. Inturn
it expects this durationInYears data info to be available
under a srelMetaData key within the gData dictionary. If
this is not the case, then it will trigger a procdata_ex
to calc the same.

Also make procdata_ex work with either a single operation,
or a list of operations.