AnalDataSimple:Roll metadata account for valid duration
As rolling return will not have any meaningful data at the begining
of the dateRange, till the rolling window is fully within the
dateRange, so ensure that we dont try to put out average rolling
rate for such period at the beginning.

Also ensure to put blank label, if there is no valid meta data.