MFExploreScript: Avoid unnecessary dma-roll normal ranking
As the latest RollingRet, in itself is not significant beyond
being what the return will be just around a short interval of
current times, so avoid ranking based on dma50Roll1095.

NOTE: It has served the purpose of testing the related logic.

Also remove some of the old and no longer used code.