Optional options argument to proc_days.
Main logic updated to expect a options argument to be passed
to proc_days and inturn the date handler passed to it.

The fetch and load data logics have been updated to expect
a optional opts argument, which inturn it will pass to the
underlying date handler called by proc_days.

indiamf being updated to account for same.

Fetch will now look for two possible options ForceLocal and
ForceRemote. Based on these options and health of data pickle
file, it will decide how to handle the fetch.