AnalDataSimple: few loaded Entities, forceValEntities
Alert the users through readme that, they have to think twice when
interpreting block_ranked sorted list, if the number of entities
loaded is small (i.e the total number of entities to work with,
not the subset that is selected through entCodes).

Renamed _clear4entities to forceval_entities to reflect its true
semantic. Also fixed its flow to reflect things better and to be
flexible.