infoset1: look at entities in a predefined way
Added a simple logic to look at available data in a predefined way
for a given set of entities.

Idea is to simplify some common ways of looking at data, into simple
commands, instead of having to key in all the required commands for
each specific way of looking at the data.