active loadFilter moved to gMeta, pickle meta
Pickle related logic updated to save and load gMeta related data,
where required. For Example required wrt save and restore session,
but not required for fetch and load related pickle (so a dummy meta
passed).

The active loadFilter match templates wrt type and name which is part
of gData now moved to gMeta.