IndiaMF: fix gMeta oversight, try fetch if load fails
Fix a oversight wrt not passing gMeta to hlpr's gdata_add logic.

Also fix the oversight wrt not referencing the global gMeta, from
within the setup of indiamf module.

Now when indiamf tries to load the data belonging to a given date
from the saved pickle for the day, if it is not successfuly, then
it will try to force a remote fetch of the data csv file.