Helper:Match templates make matchTempaltes into list if required
as match templates argument passed to matches_templates is expected
to be a list of strings, so if the user wrongly passes a single
string, then it will be automatically converted to a list of that
single string, so that the logic can work properly. This also
makes it easy for users who use matches_templates function.