IndiaMF: Allow for smaller pickle files for MF initial days
During 2006 from which the AMFI MF related data is available, one
can notice that the amount of MF data (ie number of MFs) is less,
so allow for smaller pickle files.

Also update a note wrt world.py, that it is just a simple initial
test code for now and nothing more.