LoadData: call backs; Indexes add entType during loading
Had forgotten to call the load_data callbacks belonging to the
modules, this has been fixed now.

Previously I was adding the entTYpe related to indexes, in the
setup logic of indexes module, but now instead it is done in its
load_data callback, because load_data clears the gMeta at the
beginning.

Currently the main module doesnt allow opts wrt load_data, nor does
indexes module make use of the same, but just a dummy one assigned
to None is left in indexes module.