SessionRestore: setup modules also
With the new multi module logic, one requires to setup the modules
to work with the newly restored gData/gMeta combo. This is handled
now. There could be still some additional things to save and restore,
but already most of the core logic wrt session save and restore is
taken care of.

Readme updated wrt session save/restore.