Main: Switch from current dataIndex to nxtDateIndex
The previous 'dataIndex' actually represented the dateIndex i.e the
index where data will be available for the last date in the data set.

However nxtDateIndex, will point to the index beyond the currently
last date in the data set. So to get the actual last valid index
in the dataset, which corresponds to the last date for which data
is available, one will have to -1.