DataSrc: RemoveInvalid if ForceLocal and No pickle found
If there is no valid pickle file and the logic is asked to first
check for a local copy of the remote file, now the logic rechecks
that the remote file which was previously downloaded is fine or
not.

As the same check is also required after a remote file is downloaded
for the 1st time, so the common logic is moved into a new func,
called from both locations now.