ProcData:RetOn: 7D not 5D, Off by 1
As we are looking for how a entity is performing compared to a week
back, so use 7days (i.e 5 working days roughly, ignoring holidays
for now).

As we look at things using RetOn operation, we require to look at
how entity has performed wrt 1D, 7D, 1M, ... back. So fix the
off by 1 in the logic.