Plot:Spline: Extrapolate to 2 additional days
NOTE: One will be lucky if the reality turns out anywhere near a
100 million mile of what this suggests. So dont bother.

The reason is because market reacts and moves based on realities
of the world and not based on fine grained historic pattern. The
max it may follow the historic trend in some rare cases, while
this tries to get a fine grained approximation, so chances are
will be off wrt reality by a million miles.