Plot: Use IndexFormatter to show date in plot x axis
THis allows one to zoom in or out and still get appropriate dates
shown. As well as allow one to move along x-axis and see the date
associated with the corresponding x location.

Also fix a small oversight wrt the explicit x-axis range data that
was introduced in a previous commit.