Entities: support Aliases to dataKeys, OpenHighLowCloseVolume
Added support for aliases to entities. So that one can specify
the aliases during initialisation, as well as optimize takes
care of updating the aliases.

The program now uses this aliases support to map nav and close
to data dataKey.

Nav is used by MF, while Close is used by Stocks. Both map to
dataKey called data at the lower level.

With this the changes done in the last commit wrt India module
using 'nav' and 'close' is handled as required now.

Also dataKeys open, close, high, low and volume have been added
to the main program, so that support for same added to india
stocks can be handled properly.