ProcEDB/Stocks: Use even number of days wrt Moving Avg
To keep the logic very very simple for now, see to it that number
of days is even wrt moving average operations.

TODO: May add explicit logic to convert odd number to its nearest
smaller even number and log a warning. Or on second thought leave
it has it is so that it raises a exception and user knows the
need to change num days used wrt moving average, explicitly.