Plot:Stocks: small transparent inset volume plot at bottom
start on adding a insert small plot at the bottom of the main plot,
which will contain the volume info for the stock being plotted.

TODO: YTicks need to be moved to other side
TODO: Need to make x-axis of both main and inset plot aligned and
one to one.