Ops:WeeklyView: Fix up the logic to round out the implementation
Decide the week length based on whether bSkipWeekends is enabled
or not wrt the source entities db. And inturn use the same to decide
on the lenght of the new weekly dataset.

Create the dest data keys once at the begining.

Fix the oversight wrt forgetting to update the dest data index,
as the dest data is being generated.

Import ops into the main program, so user can use the same as reqd.