OpsCU: RelTo updated to be parallel in one shot
NOTE: As of now the meta labels are explicitly created and stored
in the edb, and this inturn uses a loop through each entity. This
will be changed later, such that the labels are not stored, instead
they are created as and when required by calling into a generic
metadata to string function.

	TOwards this end is the reason why the md2str funcs are
	being defined for each op.