OpsCU: Partly Parallelise BlockStats
Note: In this case both dataDst.MetaData as well as dataDst.MetaLabel
are created serially one entity at a time. Need to see if I can create
the MetaData array without iterating in python. As for MetaLabels, they
will be avoided at a later date, once md2str demuxer to appropriate
md2str call (based on the operation which generated the metadata)
is implemented.