Ops:BlocksStat: AvgBlockAvgs and AvgBlockStds directly in MetaData
So that user can use anal_simple to sort entities using any of these.
NOTE: As currently RollAvgs looks at only valid data when averaging,
without accounting for the period when there is no data, so some
users may want to account for time when no data is available, and
in this case blockStats may be better.

	THis wont work directly because the MetaData of blockstats
	is currently objects. Need to change this later (TODO).

NOTE: Also the above behaviour of blockstats, where time with no
data is also accounted wrt mean-std-... calculation is new behaviour,
which was not there in the old code in procedb.