Entities: Switch to DateIndex wrt First/LastSeen
As in most places I require to work with dates in a relative fashion
and not the absolute dates, so storing first and last seen dates
wrt entities has a index into the dates list/array, makes it easy
to operate with them faster in most of the use cases.

So switching from storing firstSeen and lastSeen as dates to now
they being dateIndexes so also renamed to firstSeenDI/lastSeenDI

Next need to update the locations using the same.