Ops: Wider Cols and Align column headers wrt mdhdr ;(
As value of some of the entities are large, so increased teh space
alloted to values from 9 to 10 chars.

Also as some entities may have all their returns below the MinRetPA,
so the space alloted for same has been increased from 5 to 6 (100.00)

Also reverted to having % nd %pa as part of format string, now that
the argument column header names are right aligned.