ProcEDB:AnalSimple: Normal AnalType data as float
For the Normal AnalType, ensure that the data one is operating on
(as selected using theDate or theIndex) is float by explicitly
converting to float (just in case) before using it.

NOTE: This helps for example when operating on blockstats meta
data, which is stored as a array of objects (whcih inturn could
be direct floats or arrays of float or ...).