IndiaSTKs:Purpose/ActionType: Rs./Re.; Split no fv token
The Bc file from NSE has

RS.|Rs.|RE.|Re. in some cases, this is handled now. The dot i.e '.'
in this case doesnt signify rupee.paise, or rather 0.paise BUT rather
just rupee only. i.e '.' needs to be ignored in these cases and the
same is done now.

Even thou there are SPLIT corporate actions without FV tag/token in it,
it still seems to be represented in face value of share only. So it is
handled in the same way i.e SPLIT with and without FV tag are handled
the same way now.